requirements.txt
appdirs==1.4.3
asgiref==3.2.10
boto3==1.15.12
botocore==1.18.12
branca==0.4.1
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
dj-places==4.0.0
Django==3.1.1
django-allauth==0.42.0
django-cleanup==5.1.0
django-comments-xtd==2.7.2
django-contrib-comments==1.9.2
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-mapbox-location-field==1.6.3
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-storages==1.10.1
django-tinymce4-lite==1.8.0
djangorestframework==3.11.1
docutils==0.16
folium==0.11.0
future==0.18.2
geocoder==1.38.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsmin==2.2.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.19.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
ratelim==0.1.6
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
whitenoise==5.2.0
